fixed frequency int -> double in scene.h

+++ src/scene.h
@@ -61,7 +61,7 @@ public:
     QString description;
 
     // harmonic
-    int frequency;
+    double frequency;
 
     // transient
     AnalysisType analysisType;
